Unify dashboard deployment with docs deploy
