+++ .github/workflows/docs-deploy.yml
@@ -5,15 +5,16 @@ on:
     branches:
       - main
   workflow_dispatch:
+  workflow_run:
+    workflows: ["Benchmark Dispatch"]
+    types: [completed]
 
-# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
 permissions:
   contents: read
   pages: write
   id-token: write
+  actions: read
 
-# Allow only one concurrent deployment, skipping runs queued between the run in-progress and latest queued.
-# However, do NOT cancel in-progress runs as we want to allow these production deployments to complete.
 concurrency:
   group: "pages"
   cancel-in-progress: false
@@ -53,10 +54,20 @@ jobs:
           cd docs/
           make html
 
-      - name: Add dashboard
+      - name: Build dashboard data
+        env:
+          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          python3 .github/dashboard/build_dashboard_data.py \
+            --repo ${{ github.repository }} \
+            --existing-url https://helionlang.com/dashboard/dashboard-data.json \
+            --output ./dashboard-data.json
+
+      - name: Add dashboard to site
         run: |
           mkdir -p ./site/dashboard
           cp .github/dashboard/index.html ./site/dashboard/index.html
+          cp ./dashboard-data.json ./site/dashboard/dashboard-data.json
 
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v5
